First working WebUI
Added a WebUI usign Muntjac and ZeroMQ. The Webui can be disabled on the commandline or with a config file. It currently only allows buttons to be pressed that are an event (except for adding a play button as well). It's tested on Debian Stretch, and requires Python2, muntjac, zeromq in addition to the ear dependencies. Muntjac is installed on Python2 using pip. This is only an example, but might just work